fix: separate dist/ files by package name
This patch aims to prevent packages from loading unnecessary chunks by
seperating them by folders of package name
